Moved auth to be an exported variable in firebase config
