+++ auth.ts
@@ -1,6 +1,5 @@
 import { createUserWithEmailAndPassword, getAuth, signInWithEmailAndPassword, signOut } from "firebase/auth";
-
-export const auth = getAuth();
+import { auth } from './firebaseConfig';
 
 export const register = async (email: string, password: string) => {
     const userCredential = await createUserWithEmailAndPassword(auth, email, password);
